Refactoring + packages
