--- src/main/java/pl/edu/agh/iet/katabank/BankProduct.java
@@ -1,7 +1,0 @@
-package pl.edu.agh.iet.katabank;
-
-public interface BankProduct {
-
-    Customer getOwner();
-
-}
