+++ src/main/java/pl/edu/agh/iet/katabank/bankproduct/BankProduct.java
@@ -1,0 +1,9 @@
+package pl.edu.agh.iet.katabank.bankproduct;
+
+import pl.edu.agh.iet.katabank.Customer;
+
+public interface BankProduct {
+
+    Customer getOwner();
+
+}
