+++ src/main/java/pl/edu/agh/iet/katabank/repository/InMemoryBankProductsRepository.java
@@ -1,4 +1,8 @@
-package pl.edu.agh.iet.katabank;
+package pl.edu.agh.iet.katabank.repository;
+
+import pl.edu.agh.iet.katabank.Customer;
+import pl.edu.agh.iet.katabank.bankproduct.Account;
+import pl.edu.agh.iet.katabank.bankproduct.Deposit;
 
 import java.util.HashMap;
 import java.util.HashSet;
