+++ src/test/groovy/pl/edu/agh/iet/katabank/bankproduct/AccountTest.groovy
@@ -1,5 +1,7 @@
-package pl.edu.agh.iet.katabank
+package pl.edu.agh.iet.katabank.bankproduct
 
+import pl.edu.agh.iet.katabank.Customer
+import pl.edu.agh.iet.katabank.bankproduct.Account
 import spock.lang.Specification
 
 import static org.assertj.core.api.Assertions.assertThat
